🐛 fix: market popular 조회 로직 오타 수정

+++ src/main/java/no/gunbang/market/common/TradeCount.java
@@ -15,5 +15,5 @@ public class TradeCount {
     @Id
     private Long itemId;
 
-    private int count = 0;
+    private long count = 0;
 }
